seperate index and test
